chore: refactor git protect branch check
