+++ app/server/appsmith-server/src/main/java/com/appsmith/server/helpers/ce/GitPrivateRepoHelperCE.java
@@ -1,8 +1,11 @@
 package com.appsmith.server.helpers.ce;
 
+import com.appsmith.server.domains.GitApplicationMetadata;
 import reactor.core.publisher.Mono;
 
 public interface GitPrivateRepoHelperCE {
 
     Mono<Boolean> isRepoLimitReached(String workspaceId, Boolean isClearCache);
+
+    Mono<Boolean> isProtectedBranch(String branchName, GitApplicationMetadata gitApplicationMetadata);
 }
