+++ app/server/appsmith-server/src/main/java/com/appsmith/server/services/ce/GitServiceCE.java
@@ -80,6 +80,4 @@ public interface GitServiceCE {
     Mono<BranchTrackingStatus> fetchRemoteChanges(String defaultApplicationId, String branchName, boolean isFileLock);
 
     Mono<String> autoCommitDSLMigration(String defaultApplicationId, String branchName);
-
-    Mono<Boolean> isProtectedBranch(String branchName, GitApplicationMetadata gitApplicationMetadata);
 }
