+++ app/server/appsmith-server/src/main/java/com/appsmith/server/services/ce/GitServiceCEImpl.java
@@ -427,7 +427,8 @@ public class GitServiceCEImpl implements GitServiceCE {
 
         boolean isSystemGenerated = isSystemGeneratedTemp;
         Mono<String> commitMono = this.getApplicationById(defaultApplicationId)
-                .flatMap(application -> isProtectedBranch(branchName, application.getGitApplicationMetadata())
+                .flatMap(application -> gitPrivateRepoHelper
+                        .isProtectedBranch(branchName, application.getGitApplicationMetadata())
                         .flatMap(status -> {
                             if (Boolean.TRUE.equals(status)) {
                                 return Mono.error(new AppsmithException(
@@ -1597,7 +1598,8 @@ public class GitServiceCEImpl implements GitServiceCE {
          * */
 
         Mono<GitPullDTO> pullMono = getApplicationById(defaultApplicationId)
-                .flatMap(application -> isProtectedBranch(branchName, application.getGitApplicationMetadata())
+                .flatMap(application -> gitPrivateRepoHelper
+                        .isProtectedBranch(branchName, application.getGitApplicationMetadata())
                         .flatMap(status -> {
                             if (Boolean.TRUE.equals(status)) {
                                 return Mono.error(new AppsmithException(
@@ -2704,7 +2706,8 @@ public class GitServiceCEImpl implements GitServiceCE {
     @Override
     public Mono<Application> deleteBranch(String defaultApplicationId, String branchName) {
         Mono<Application> deleteBranchMono = getApplicationById(defaultApplicationId)
-                .flatMap(application -> isProtectedBranch(branchName, application.getGitApplicationMetadata())
+                .flatMap(application -> gitPrivateRepoHelper
+                        .isProtectedBranch(branchName, application.getGitApplicationMetadata())
                         .flatMap(status -> {
                             if (Boolean.TRUE.equals(status)) {
                                 return Mono.error(new AppsmithException(
@@ -2953,11 +2956,6 @@ public class GitServiceCEImpl implements GitServiceCE {
                                     }))
                             .flatMap(pushResult -> pushApplicationErrorRecovery(pushResult, application)));
         });
-    }
-
-    @Override
-    public Mono<Boolean> isProtectedBranch(String branchName, GitApplicationMetadata gitApplicationMetadata) {
-        return Mono.just(Boolean.FALSE);
     }
 
     private Mono<Application> deleteApplicationCreatedFromGitImport(
